# api/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "The Portal Messaging API.",
        "version": "0.0.1"
    },
    "host": "dev.portalmessaging.com",
    "basePath": "/v1",
    "paths": {
...
        "/login/google": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login or register via a Google account.",
                "operationId": "googleLogin",
                "parameters": [
                    {
                        "name": "google_login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/googleLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/loginResponse"
                    },
                    "400": {
                        "$ref": "#/responses/error"
                    },
                    "500": {
                        "$ref": "#/responses/error"
                    },
                    "default": {
                        "$ref": "#/responses/detailError"
                    }
                }
            }
        },
...
        "/user/devices": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "schemes": [
                    "https"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Retrieve a user's existing connected devices.",
                "operationId": "getDevices",
                "parameters": [
                    {
                        "type": "string",
                        "name": "X-USER-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "X-USER-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/deviceList"
                    },
                    "401": {
                        "$ref": "#/responses/error"
                    },
                    "500": {
                        "$ref": "#/responses/error"
                    },
                    "default": {
                        "$ref": "#/responses/detailError"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "schemes": [
                    "https"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register a new Google Cloud Messaging device.",
                "operationId": "addDevice",
                "parameters": [
                    {
                        "type": "string",
                        "name": "X-USER-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "X-USER-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "add_device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/addDevice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/addDevice"
                    },
                    "401": {
                        "$ref": "#/responses/error"
                    },
                    "500": {
                        "$ref": "#/responses/error"
                    },
                    "default": {
                        "$ref": "#/responses/detailError"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "addDevice": {
            "type": "object",
            "required": [
                "registration_id",
                "name",
                "type"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "registration_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "pattern": "(phone,chrome,desktop)"
                }
            }
        },
        "addDeviceResponse": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "encryption_key": {
                    "type": "string"
                },
                "notification_key": {
                    "type": "string"
                }
            }
        },
        "detailError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "deviceListResponse": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/linkedDevice"
                    }
                }
            }
        },
        "errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "googleLogin": {
            "type": "object",
            "required": [
                "id_token"
            ],
            "properties": {
                "id_token": {
                    "type": "string"
                }
            }
        },
        "linkedDevice": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "loginResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                },
                "user_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "responses": {
        "addDevice": {
            "description": "AddDeviceResponse contains the encryption and\nnotifica†ion keys for a new GCM client.",
            "schema": {
                "$ref": "#/definitions/addDeviceResponse"
            }
        },
        "detailError": {
            "description": "A DetailError gives both the error and reason for the error.",
            "schema": {
                "$ref": "#/definitions/detailError"
            }
        },
        "deviceList": {
            "description": "DeviceListResponse contains existing, connected user devices.",
            "schema": {
                "$ref": "#/definitions/deviceListResponse"
            }
        },
        "error": {
            "description": "An ErrorResponse wraps a bad request with an error code string.",
            "schema": {
                "$ref": "#/definitions/errorResponse"
            }
        },
        "loginResponse": {
            "description": "A LoginResponse contains credentials to make authenticated requests.",
            "schema": {
                "$ref": "#/definitions/loginResponse"
            }
        },
...
    }
}